# contracts/openapi/openapi.yaml
openapi: 3.1.1
info:
  title: API SysAP
  version: 0.2.0
  description: >-
    Contrato HTTP do SysAP. Saude e prontidao estao implementadas; a superficie
    versionada de identidade descreve o contrato planejado para a Fase 2 e nao
    representa funcionalidade disponivel nesta subfase documental.
servers:
  - url: http://127.0.0.1:8080
    description: API local em loopback
security:
  - BearerAccessToken: []
...
paths:
...
  /v1/activation/start:
    post:
      operationId: startActivation
      tags:
        - Convites de atleta
      summary: Iniciar ativação de atleta
      description: >-
        Inicia o fluxo de ativação, gerando um OTP para a matrícula informada.
        Sempre retorna 202 Accepted para não enumerar matrículas válidas.
        Em desenvolvimento, o provedor é local; em produção, o envio é
        configurado no Supabase Auth. A API não integra provedores SMS diretamente.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationStartRequest"
      responses:
        "202":
          description: Solicitação aceita
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: accepted
        "503":
          $ref: "#/components/responses/IdentityVerificationUnavailable"
  /v1/activation/verify:
    post:
      operationId: verifyActivation
      tags:
        - Convites de atleta
      summary: Verificar OTP
      description: >-
        Verifica se o OTP fornecido corresponde à matrícula e não expirou.
        Retorna uma prova de ativação (activation_proof) que pode ser usada depois.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationVerifyRequest"
      responses:
        "200":
          description: OTP validado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationVerifyResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "503":
          $ref: "#/components/responses/IdentityVerificationUnavailable"
  /v1/activation/complete:
    post:
      operationId: completeActivation
      tags:
        - Convites de atleta
      summary: Concluir ativação
      description: >-
        Conclui o fluxo de ativação, validando a activation_proof e definindo a senha do atleta.
        Provisiona a identidade pelo adaptador server-side do Supabase Auth.
      x-sysap-implementation-status: implemented
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationCompleteRequest"
      responses:
        "200":
          description: Ativação concluída com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationCompleteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "409":
          $ref: "#/components/responses/ResourceConflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    BearerAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Access token de curta duracao validado pela API. O token e tratado pelo
        cliente como opaco; sub/session_id nao sao IDs de negocio. Sessao deve
        ter sido registrada pela API; papel, organizacao e acesso sao
        confirmados no estado do SysAP.
...
  responses:
...
    BadRequest:
      description: Requisicao malformada ou header invalido.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            malformedRequest:
              summary: Requisicao malformada
              value:
                error:
                  code: invalid_request
                  message: request is invalid
                  request_id: 0123456789abcdef0123456789abcdef
...
    AuthenticationRequired:
      description: >-
        Access token ausente, invalido ou expirado, ou sessao local inexistente
        ou revogada.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        WWW-Authenticate:
          $ref: "#/components/headers/BearerChallenge"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            authenticationRequired:
              summary: Autenticacao necessaria
              value:
                error:
                  code: authentication_required
                  message: authentication is required
                  request_id: 0123456789abcdef0123456789abcdef
...
    IdentityVerificationUnavailable:
      description: >-
        Nenhuma chave JWKS confiavel esta disponivel para verificacao de
        identidade. Esta resposta e planejada para GET /v1/me e nao revela
        detalhes do token, claim, emissor ou infraestrutura.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            identityVerificationUnavailable:
              summary: Verificacao de identidade indisponivel
              value:
                error:
                  code: identity_verification_unavailable
                  message: identity verification is temporarily unavailable
                  request_id: 0123456789abcdef0123456789abcdef
...
    ResourceConflict:
      description: Estado incompativel ou chave de idempotencia reutilizada.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            resourceConflict:
              summary: Conflito de estado
              value:
                error:
                  code: resource_conflict
                  message: request conflicts with current state
                  request_id: 0123456789abcdef0123456789abcdef
    ValidationFailed:
      description: Campos invalidos sem ecoar valores sensiveis recebidos.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          examples:
            validationFailed:
              summary: Falha de validacao estatica
              value:
                error:
                  code: validation_failed
                  message: request validation failed
                  request_id: 0123456789abcdef0123456789abcdef
                  details:
                    - field: enrollment_number
                      reason: invalid_format
    RateLimited:
      description: Limite temporario excedido sem revelar a dimensao aplicada.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            rateLimited:
              summary: Limite temporario excedido
              value:
                error:
                  code: rate_limited
                  message: too many requests
                  request_id: 0123456789abcdef0123456789abcdef
    ServiceUnavailable:
      description: >-
        Operacao temporariamente indisponivel sem identificar dependencia ou
        fornecedor. Em fluxos antienumeracao, somente falha global usa 503.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            serviceUnavailable:
              summary: Servico temporariamente indisponivel
              value:
                error:
                  code: service_unavailable
                  message: service is temporarily unavailable
                  request_id: 0123456789abcdef0123456789abcdef
  headers:
...
    RequestID:
      description: >-
        Identificador hexadecimal gerado pelo servidor para correlacao da
        requisicao. O caminho normal possui 32 caracteres; a contingencia pode
        produzir um valor hexadecimal maior.
      required: true
      schema:
        $ref: "#/components/schemas/RequestID"
      example: 0123456789abcdef0123456789abcdef
    NoStore:
      description: Impede armazenamento de respostas de identidade e autenticacao.
      required: true
      schema:
        type: string
        const: no-store
      example: no-store
    RetryAfter:
      description: Segundos inteiros antes de uma nova tentativa permitida.
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 86400
      example: 60
    BearerChallenge:
      description: Desafio padrao para access token bearer.
      required: true
      schema:
        type: string
        const: Bearer
      example: Bearer
...
  schemas:
    RequestID:
      type: string
      pattern: "^[0-9a-f]+$"
      minLength: 32
      description: Identificador hexadecimal opaco gerado pelo servidor.
...
    EnrollmentNumber:
      type: string
      minLength: 10
      maxLength: 10
      pattern: "^[0-9]{10}$"
      description: Matricula publica ficticia de exatamente dez digitos.
      example: "2026000001"
...
    OTPInput:
      type: string
      minLength: 6
      maxLength: 6
      pattern: "^[0-9]{6}$"
      writeOnly: true
      description: OTP numerico de seis digitos, tratado como segredo de uso unico.
...
    ActivationStartRequest:
      type: object
      additionalProperties: false
      required:
        - enrollment_number
      properties:
        enrollment_number:
          $ref: "#/components/schemas/EnrollmentNumber"
    ActivationVerifyRequest:
      type: object
      additionalProperties: false
      required:
        - enrollment_number
        - code
      properties:
        enrollment_number:
          $ref: "#/components/schemas/EnrollmentNumber"
        code:
          $ref: "#/components/schemas/OTPInput"
    ActivationVerifyResponse:
      type: object
      additionalProperties: false
      required:
        - activation_proof
      properties:
        activation_proof:
          type: string
          description: Prova de validação do OTP (hexadecimal, única, opaca)
          example: 1a2b3c4d5e6f7g8h
    ActivationCompleteRequest:
      type: object
      required:
        - activation_proof
        - password
      properties:
        activation_proof:
          type: string
          description: Prova obtida no step de verify
          example: 1a2b3c4d5e6f7g8h
        password:
          type: string
          format: password
          description: Senha do atleta
    ActivationCompleteResponse:
      type: object
      properties:
        profile_id:
          type: string
          format: uuid
        enrollment_number:
          type: string
        role:
          type: string
          example: athlete
        organization_id:
          type: string
          format: uuid
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/PublicError"
    PublicError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          minLength: 3
          maxLength: 64
          pattern: "^[a-z][a-z0-9_]*$"
        message:
          type: string
          minLength: 1
          maxLength: 200
        request_id:
          $ref: "#/components/schemas/RequestID"
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - request_id
        - details
      properties:
        code:
          type: string
          const: validation_failed
        message:
          type: string
          const: request validation failed
        request_id:
          $ref: "#/components/schemas/RequestID"
        details:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/ValidationDetail"
    ValidationDetail:
      type: object
      additionalProperties: false
      required:
        - field
        - reason
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[a-z][a-z0-9_.]*$"
        reason:
          type: string
          enum:
            - required
            - invalid_format
            - too_short
            - too_long
            - not_allowed
            - password_rejected